Remove DeleteByIdTask indirection
This inlines the logic into the DeleteById plan. The separate task is a
leftover from when we still had a Task interface for the execution
logic.
